# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AGA Health Foundation Telehealth Readiness Survey API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prospectus:
    post:
      operationId: createProspectus
      tags: [prospectus]
      summary: Create draft prospectus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProspectusInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusPublic"
  /prospectus/{publicId}:
    get:
      operationId: getProspectus
      tags: [prospectus]
      summary: Get prospectus by public token
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Prospectus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusPublic"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchProspectus
      tags: [prospectus]
      summary: Update draft or revision-requested prospectus
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchProspectusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusPublic"
  /prospectus/{publicId}/submit:
    post:
      operationId: submitProspectus
      tags: [prospectus]
      summary: Submit prospectus for review
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusPublic"
  /prospectus/{publicId}/withdraw:
    post:
      operationId: withdrawProspectus
      tags: [prospectus]
      summary: Withdraw prospectus
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusPublic"
  /prospectus/{publicId}/attachments:
    post:
      operationId: uploadProspectusAttachment
      tags: [prospectus]
      summary: Upload supporting document (Vercel Blob)
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectusAttachmentUploadInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProspectusAttachmentMeta"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProspectusStatus:
      type: string
      enum:
        - draft
        - submitted
        - under_review
        - revision_requested
        - approved
        - rejected
        - withdrawn
    ProspectusApprovalRole:
      type: string
      enum:
        - research_leadership
        - platform_ops
    CreateProspectusInput:
      type: object
      required: [submitterEmail, submitterName]
      properties:
        submitterEmail:
          type: string
          format: email
        submitterName:
          type: string
        title:
          type: string
        principalInvestigator:
          type: string
    PatchProspectusInput:
      type: object
      properties:
        title:
          type: string
        researchProblem:
          type: string
        proposedSlug:
          type: string
        studyTemplate:
          type: string
          enum: [telehealth-readiness-clone, clinician-clone, custom]
    ProspectusApprovalSlot:
      type: object
      required: [role, decision, comments, decidedAt]
      properties:
        role:
          $ref: "#/components/schemas/ProspectusApprovalRole"
        decision:
          type: ["string", "null"]
          enum: [approved, rejected, null]
        comments:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
          format: date-time
    ProspectusPublic:
      type: object
      required: [publicId, status, submitterEmail, submitterName]
      properties:
        id:
          type: integer
        publicId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProspectusStatus"
        submitterEmail:
          type: string
        submitterName:
          type: string
        title:
          type: string
        principalInvestigator:
          type: string
        proposedSlug:
          type: ["string", "null"]
        studyTemplate:
          type: string
        linkedStudySlug:
          type: ["string", "null"]
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/ProspectusApprovalSlot"
        attachmentCount:
          type: integer
...
    ProspectusAttachmentUploadInput:
      type: object
      required: [filename, mimeType, contentBase64]
      properties:
        filename:
          type: string
        mimeType:
          type: string
        contentBase64:
          type: string
    ProspectusAttachmentMeta:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        uploadedAt:
          type: string
          format: date-time